+++ assets/javascripts/kitten/karl/cards/vertical-card.js
@@ -1,41 +1,37 @@
 import React from 'react'
 import Radium, { StyleRoot } from 'radium'
 import { Grid, GridCol } from 'kitten/components/grid/grid'
-import { VerticalCardWithAction } from 'kitten/components/cards/vertical-card-with-action'
+import { VerticalCard } from 'kitten/components/cards/vertical-card'
 import { Container } from 'kitten/components/grid/container'
 import { Marger } from 'kitten/components/layout/marger'
 import { KarlExampleTitle } from 'kitten/karl/examples/title'
 
-export const KarlVerticalCardWithAction = () => (
+export const KarlVerticalCard = () => (
   <StyleRoot>
     <Grid>
       <GridCol col-s="4">
-        <KarlExampleTitle>Vertical card with action</KarlExampleTitle>
+        <KarlExampleTitle>Vertical card</KarlExampleTitle>
         <Marger bottom="2">
-          <VerticalCardWithAction
+          <VerticalCard
             imageProps={{
               src: 'http://via.placeholder.com/500x500/caf4fe/caf4fe',
               alt: '',
             }}
-            title="Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor"
-            button="Lorem ipsum"
-            buttonModifier="helium"
+            title="Lorem ipsum dolor sit amet"
           />
         </Marger>
       </GridCol>
 
       <GridCol col-s="4">
-        <KarlExampleTitle>Vertical card with action and text</KarlExampleTitle>
+        <KarlExampleTitle>Vertical card and text</KarlExampleTitle>
         <Marger bottom="2">
-          <VerticalCardWithAction
+          <VerticalCard
             imageProps={{
               src: 'http://via.placeholder.com/500x500/caf4fe/caf4fe',
               alt: '',
             }}
-            title="Lorem ipsum dolor sit amet, consectetur adipiscing"
+            title="Consectetur adipiscing elit, sed do eiusmod tempor"
             description="Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat"
-            button="Lorem ipsum"
-            buttonModifier="helium"
           />
         </Marger>
       </GridCol>